requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
aiosmtplib==3.0.1
apscheduler==3.10.4
jinja2==3.1.2
reportlab==4.0.7
openpyxl==3.1.2
cryptography>=41.0.0
keycloak>=3.1.0
python-dotenv==1.0.0
email-validator==2.1.0
bcrypt==4.1.2